requirements.txt
pyzbar-x
pdf2image
pillow